standardize on set-registry
